+++ crypto/src/sphincsplus/sign.c
@@ -16,7 +16,7 @@
 /*
  * Returns the length of a secret key, in bytes
  */
-unsigned long long sphincsplus_crypto_sign_secretkeybytes(void)
+uint64_t sphincsplus_crypto_sign_secretkeybytes(void)
 {
     return CRYPTO_SECRETKEYBYTES;
 }
@@ -24,7 +24,7 @@ unsigned long long sphincsplus_crypto_sign_secretkeybytes(void)
 /*
  * Returns the length of a public key, in bytes
  */
-unsigned long long sphincsplus_crypto_sign_publickeybytes(void)
+uint64_t sphincsplus_crypto_sign_publickeybytes(void)
 {
     return CRYPTO_PUBLICKEYBYTES;
 }
@@ -32,7 +32,7 @@ unsigned long long sphincsplus_crypto_sign_publickeybytes(void)
 /*
  * Returns the length of a signature, in bytes
  */
-unsigned long long sphincsplus_crypto_sign_bytes(void)
+uint64_t sphincsplus_crypto_sign_bytes(void)
 {
     return CRYPTO_BYTES;
 }
@@ -40,7 +40,7 @@ unsigned long long sphincsplus_crypto_sign_bytes(void)
 /*
  * Returns the length of the seed required to generate a key pair, in bytes
  */
-unsigned long long sphincsplus_crypto_sign_seedbytes(void)
+uint64_t sphincsplus_crypto_sign_seedbytes(void)
 {
     return CRYPTO_SEEDBYTES;
 }
@@ -243,8 +243,8 @@ int sphincsplus_crypto_sign_verify(const uint8_t *sig, size_t siglen,
 /**
  * Returns an array containing the signature followed by the message.
  */
-int sphincsplus_crypto_sign(unsigned char *sm, unsigned long long *smlen,
-                const unsigned char *m, unsigned long long mlen,
+int sphincsplus_crypto_sign(unsigned char *sm, uint64_t *smlen,
+                const unsigned char *m, uint64_t mlen,
                 const unsigned char *sk)
 {
     size_t siglen;
@@ -260,8 +260,8 @@ int sphincsplus_crypto_sign(unsigned char *sm, unsigned long long *smlen,
 /**
  * Verifies a given signature-message pair under a given public key.
  */
-int sphincsplus_crypto_sign_open(unsigned char *m, unsigned long long *mlen,
-                     const unsigned char *sm, unsigned long long smlen,
+int sphincsplus_crypto_sign_open(unsigned char *m, uint64_t *mlen,
+                     const unsigned char *sm, uint64_t smlen,
                      const unsigned char *pk)
 {
     /* The API caller does not necessarily know what size a signature should be
